requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]